tests(api-core): move pools to test-support mod
