+++ crates/api-core/src/test_support/ib_guid_pool.rs
