+++ crates/api-core/src/tests/machine_boot_override.rs
@@ -17,10 +17,10 @@
 
 use common::api_fixtures::create_test_env;
 use common::api_fixtures::dpu::dpu_discover_dhcp;
-use common::mac_address_pool::DPU_OOB_MAC_ADDRESS_POOL;
 use rpc::protos::forge::forge_server::Forge;
 
 use crate::DatabaseError;
+use crate::test_support::mac_address_pool::DPU_OOB_MAC_ADDRESS_POOL;
 use crate::tests::common;
 
 #[crate::sqlx_test]
